Add more fields to detail

+++ app/src/main/java/com/xteam/sonytakehome/model/BusinessDetail.kt
@@ -65,7 +65,7 @@ data class BusinessDetail(
     val display_phone: String,
     val review_count: Int,
     val categories: List<Categories>?,
-    val rating: Double,
+    val rating: Float,
     val location: Location,
     val coordinates: Coordinates,
     val photos: List<String>?,
